requirements.txt
fastapi
uvicorn[standard]
python-multipart
opencv-python
numpy
pandas
psycopg2-binary
deepface
tf-keras
